UI: Add support for "titleHtml" to configure the title/header content
In the UI configuration, add support for "titleHtml" to customize the
HTML content of the header that is displayed above the input field in a
new chat as long as no message has been sent. This is similar to "title"
which only allows a text to be set. The "titleHtml" configuration could
be used to display a custom SVG image instead of text, or in addition to
it.